Migrate to Single Activity with Nav Component
UI logic moved from MainActivity to CharacterListFragment and moved layout too
